+++ phoenix/src/drivers/sd.rs
@@ -5,6 +5,31 @@ use embassy_time::Delay;
 use embedded_hal_1::delay::DelayNs;
 use embedded_hal_bus::spi::RefCellDevice;
 use embedded_sdmmc::{Mode, SdCard, VolumeIdx, VolumeManager};
+use core::marker::PhantomData;
+
+/// Minimal time source for embedded-sdmmc
+pub struct TimeSink {
+    pub(crate) _marker: PhantomData<*const ()>,
+}
+
+impl TimeSink {
+    pub fn new() -> Self {
+        Self { _marker: PhantomData }
+    }
+}
+
+impl embedded_sdmmc::TimeSource for TimeSink {
+    fn get_timestamp(&self) -> embedded_sdmmc::Timestamp {
+        embedded_sdmmc::Timestamp {
+            year_since_1970: 0,
+            zero_indexed_month: 0,
+            zero_indexed_day: 0,
+            hours: 0,
+            minutes: 0,
+            seconds: 0,
+        }
+    }
+}
 
 // Initializes the SD card and performs the small demo previously in main.rs.
 // Returns Ok(()) if initialization and simple access succeed.
@@ -18,9 +43,7 @@ pub fn init_and_demo(
     info!("Card size is {} bytes", sdcard.num_bytes().map_err(|_| ())?);
     let volume_mgr = VolumeManager::new(
         sdcard,
-        super::super::TimeSink {
-            _marker: core::marker::PhantomData,
-        },
+        TimeSink::new(),
     );
     let _volume0 = volume_mgr.open_volume(VolumeIdx(0)).map_err(|_| ())?;
     // let root_dir = volume0.open_root_dir().map_err(|_| ())?;
