+++ phoenix/src/features.rs
@@ -3,6 +3,7 @@
 
 pub const ENABLE_LED: bool = false;
 pub const ENABLE_SBG: bool = false;
+// TODO: remove/condense these GPS features
 pub const ENABLE_GPS: bool = false;
 pub const ENABLE_GPS_POLL: bool = false;
 pub const ENABLE_GPS_PARSE: bool = false;
@@ -13,3 +14,4 @@ pub const ENABLE_BARO: bool = false;
 pub const ENABLE_RADIO: bool = true;
 pub const ENABLE_INFERENCE: bool = false;
 pub const ENABLE_SD: bool = false;
+pub const ENABLE_STATE_MACHINE: bool = true;
